refactor(arithmetic): partial fixes for some integer arithmetic lints
