+++ zebra-chain/src/block/height.rs
@@ -125,19 +125,6 @@ impl Sub<i32> for Height {
     }
 }
 
-#[cfg(any(test, feature = "proptest-impl"))]
-use proptest::prelude::*;
-#[cfg(any(test, feature = "proptest-impl"))]
-impl Arbitrary for Height {
-    type Parameters = ();
-
-    fn arbitrary_with(_args: ()) -> Self::Strategy {
-        (Height::MIN.0..=Height::MAX.0).prop_map(Height).boxed()
-    }
-
-    type Strategy = BoxedStrategy<Self>;
-}
-
 #[test]
 fn operator_tests() {
     zebra_test::init();
